Adding 400 on callback if the rack session does not have the Request Token and Request Secret that it should.
This entailed a lot of session reworking in tests. Turns out, you can
just pass 'rack.session' => { ... } into a get() (or put(), or etc
etc) call on MockRequest. This makes it super easy for us to tests
invalid sessions, etc.

Just have to remember to pass in an empty array
for the login path, because it expects a Rack::Session to set
rack.session in production.